requirements.txt
pymupdf
chromadb
langchain-community
pypdf
sentence-transformers
groq
python-dotenv
langchain-text-splitters
twilio
python-multipart
elevenlabs
httpx